Move cache implementations to internal package
