+++ instrumentation-api/src/main/java/io/opentelemetry/instrumentation/api/cache/internal/concurrentlinkedhashmap/EvictionListener.java
@@ -20,7 +20,7 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.instrumentation.api.cache.concurrentlinkedhashmap;
+package io.opentelemetry.instrumentation.api.cache.internal.concurrentlinkedhashmap;
 
 import javax.annotation.concurrent.ThreadSafe;
